requirements.txt
beautifulsoup4
requests
urllib3
mysql-connector-python